feat: app is running, need to make navigation to the paciente edit screen and the loads of insert and deletes

+++ lib/model/afericoes/frequencia_respiratoria.dart
@@ -3,7 +3,7 @@ class FrequenciaRespiratoria {
   int id;
   int idPaciente;
   int createAt;
-  int frequencia;
+  double frequencia;
 
   FrequenciaRespiratoria(
       {required this.id,
